improves pdsp::WaveTable

+++ example-wavesynth-pc-keys/src/synth.h
@@ -55,7 +55,7 @@ struct SynthGlobal {
     ofxPDSPValue     chorus_speed_ctrl;
     ofxPDSPValue     chorus_depth_ctrl;
     
-    vector<double> partials_vector;
+    vector<float> partials_vector;
 
 };
